finish crud

+++ database/migrations/2024_07_18_131842_create_data_pasiens_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('aktivitas_id')->constrained('aktivitas')->onDelete('restrict')->onUpdate('cascade');
             $table->tinyInteger('tb');
             $table->tinyInteger('bb');
-            $table->tinyInteger('kolesterol');
+            $table->integer('kolesterol');
             $table->timestamps();
         });
     }
